Feat: "TestService test"
1. TestAddRequestDto를 상속받은 테스트클래스 생성
- TestAddRequestDtoCorrect : 정상적인 TestAddRequestDto
- TestAddRequestDtoNoNickname : 닉네임이 없는 TestAddRequestDto
- TestAddRequestDtoWrongQuestions : 이상한 질문이 포함된 TestAddRequestDto

2. 클래스 Test -> Tests 이름변경

3. Validator를 이용하여 예외처리

4. addTest() 테스트 코드
- 정상적인 테스트 생성시
- 닉네임 없이 테스트를 만들때 예외 발생
- 잘못된 질문지로 테스트를 만들때 예외 발생

5. findTest() 테스트 코드
- 정상적인 테스트 조회
- 존재하지 않는 테스트를 조회할때 예외발생

6. gradeTest() 테스트 코드
- 정상적인 테스트 채점시
- 테스트의 답의 개수가 다를때의 예외발생
- 존재하지 않는 테스트를 채점 할때 예외발생

7. removeTest() 테스트 코드
- 정상적인 테스트 제거
- 존재하지 않는 테스트를 제거하려 할때 예외 발생

+++ src/main/java/com/jinbkim/whoru/tests/service/TestService.java
@@ -4,13 +4,13 @@ import com.jinbkim.whoru.questions.domain.question.Question;
 import com.jinbkim.whoru.questions.repository.QuestionRepository;
 import com.jinbkim.whoru.questions.service.QuestionService;
 import com.jinbkim.whoru.questions.web.dto.QuestionDto;
-import com.jinbkim.whoru.tests.domain.Test;
+import com.jinbkim.whoru.tests.domain.Tests;
 import com.jinbkim.whoru.tests.repository.TestRepository;
 import com.jinbkim.whoru.tests.web.dto.*;
 import java.util.stream.Collectors;
+import javax.validation.Validation;
+import javax.validation.Validator;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestParam;
 
 import java.util.*;
 
@@ -19,6 +19,7 @@ public class TestService {
     private final TestRepository testRepository;
     private final QuestionRepository questionRepository;
     private final QuestionService questionService;
+    Validator validator = Validation.buildDefaultValidatorFactory().getValidator();
 
     public TestService(TestRepository testRepository, QuestionRepository questionRepository, QuestionService questionService) {
         this.testRepository = testRepository;
@@ -26,30 +27,35 @@ public class TestService {
         this.questionService = questionService;
     }
 
-    public TestAddResponseDto addTest(@RequestBody TestAddRequestDto testAddRequestDto){
+    public TestAddResponseDto addTest(TestAddRequestDto testAddRequestDto){
         // questId 구하기
         List<String> questionIds = new ArrayList<>();
         List<QuestionDto> questions = testAddRequestDto.getQuestions();
         for(QuestionDto questionDto: questions)
             questionIds.add(questionService.addQuestion(questionDto));
 
-        // db에 test 저장
-        Test test = Test.builder()
+        // db에 tests 저장
+        Tests tests = Tests.builder()
                 .nickname(testAddRequestDto.getNickname())
                 .questionIds(questionIds)
                 .build();
-        testRepository.save(test);
+        if (validator.validate(tests).size() > 0)
+            throw new NullPointerException();
+        testRepository.save(tests);
 
         // testId 반환
-        return new TestAddResponseDto(test.getId());
+        return new TestAddResponseDto(tests.getId());
     }
 
     public TestFindResponseDto findTest(String testId) {
         // testId로 테스트의 questionIds 조회
-        Test test = testRepository.findById(testId).get();
+        Optional<Tests> testsOptional = testRepository.findById(testId);
+        if (!testsOptional.isPresent())
+            throw new NullPointerException();
+        Tests tests = testsOptional.get();
 
         // questionId로 questions 조회
-        List<QuestionDto> questionDtoList = test.getQuestionIds().stream()
+        List<QuestionDto> questionDtoList = tests.getQuestionIds().stream()
             .map(questionRepository::findById)
             .map(Optional::get)
             .map(this::createQuestionDto)
@@ -68,15 +74,20 @@ public class TestService {
             .build();
     }
 
-    public TestGradeResponseDto gradeTest(@RequestBody TestGradeRequestDto testGradeRequestDto) {
-        // testId로 test 조회
-        Test test = testRepository.findById(testGradeRequestDto.getTestId()).get();
+    public TestGradeResponseDto gradeTest(TestGradeRequestDto testGradeRequestDto) {
+        // testId로 tests 조회
+        Optional<Tests> testsOptional = testRepository.findById(testGradeRequestDto.getTestId());
+        if (!testsOptional.isPresent())
+            throw new NullPointerException();
+        Tests tests = testsOptional.get();
 
         // 정답 수 세기
         int answerCount = 0;
         List<Object> answerSubmit = testGradeRequestDto.getAnswerSubmit();
+        if (answerSubmit.size() != tests.getQuestionIds().size())
+            throw new ArrayIndexOutOfBoundsException();
         for(int i=0; i<answerSubmit.size(); i++) {
-            Question question = questionRepository.findById(test.getQuestionIds().get(i)).get();
+            Question question = questionRepository.findById(tests.getQuestionIds().get(i)).get();
             if (questionService.gradeQuestion(question.getType(), answerSubmit.get(i), question.getAnswer()))  // 정답이면
                 answerCount++;
         }
@@ -88,16 +99,19 @@ public class TestService {
                 .build();
     }
 
-    public void removeTest(@RequestParam String testId) {
-        // testId로 test 조회
-        Test test = testRepository.findById(testId).get();
+    public void removeTest(String testId) {
+        // testId로 tests 조회
+        Optional<Tests> testsOptional = testRepository.findById(testId);
+        if (!testsOptional.isPresent())
+            throw new NullPointerException();
+        Tests tests = testsOptional.get();
 
         // test의 모든 question 제거
-        List<String> questionIds = test.getQuestionIds();
+        List<String> questionIds = tests.getQuestionIds();
         for (String questionId : questionIds)
             questionRepository.delete(questionRepository.findById(questionId).get());
 
-        // test 제거
-        testRepository.delete(test);
+        // tests 제거
+        testRepository.delete(tests);
     }
 }
